requirements.txt
discord.py
discord.py[voice]
google
selenium
youtube_dl
PyNaCl